Light application fix.
Method sets uniform returns an error if uniform is not found by its
name.

Normal vector in `light` shader is normalized. It's an advice from the
book.

Ambient light is a little bit harder.

Bald cube shape is replaced by Normalized. It was the cause of the light
bug. The bald cube doesn't have data of the normals which are required
to draw diffuse light properly.

Ground is turned by one degree.

Drawing of the cubes is switched from glDrawElements to glDrawArrays.